新增主畫面閃電效果
在Scenes/UI
