+++ Assets/Scripts/MainUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b172305a26d304e4cb398b0b0097617e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
