--- Assets/Scripts/MapCreat.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 08b9332a01800e843a3c88c68177a1d6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
