--- Assets/Scripts/MapMine.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b3135b7198e305b42a384e930c9647ba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
