+++ Assets/Shader.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d1be0cab974fe05459ced78be99d46b5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
