+++ Assets/Shader/ShaderDevelopmentUsingUnity.shader
@@ -1,0 +1,62 @@
+Shader "ShaderDevelopmentUsingUnity/GradientShader"
+{
+	Properties
+	{
+		_MainTex("Texture", 2D) = "white" {}
+		_Float("_SetA", Range(0 , 1)) = 1
+	}
+		SubShader
+	{
+		Tags { "RenderType" = "Opaque"  "Queue" = "Transparent"}
+		LOD 100
+
+		Pass
+		{
+			ZWrite Off
+			Blend SrcAlpha OneMinusSrcAlpha
+
+			CGPROGRAM
+			#pragma vertex vert
+			#pragma fragment frag
+		// make fog work
+		#pragma multi_compile_fog
+
+		#include "UnityCG.cginc"
+
+		struct appdata
+		{
+			float4 vertex : POSITION;
+			float2 uv : TEXCOORD0;
+		};
+
+		struct v2f
+		{
+			float2 uv : TEXCOORD0;
+			float4 vertex : SV_POSITION;
+		};
+
+		sampler2D _MainTex;
+		float4 _MainTex_ST;
+		float	_Float;
+		v2f vert(appdata v)
+		{
+			v2f o;
+			o.vertex = UnityObjectToClipPos(v.vertex);
+			o.uv = TRANSFORM_TEX(v.uv, _MainTex);
+			UNITY_TRANSFER_FOG(o,o.vertex);
+			return o;
+		}
+
+
+		fixed4 frag(v2f i) : SV_Target
+		{
+			// sample the texture
+			fixed4 col = tex2D(_MainTex, i.uv);
+			col.a = _Float;
+
+			return col;
+		}
+		ENDCG
+	}
+	}
+}
