+++ Assets/Shader/ShaderDevelopmentUsingUnity.shader.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: d0140513ec8188e4b8e4e8763be1037c
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  preprocessorOverride: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
