+++ Assets/Shader/UI.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 71a5628ec2d25d7418234dc05cacbed7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
