+++ Assets/Test/Fire_cen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 23160e6e49caecc4dbad6692148456d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
