+++ Assets/Test/Lamp.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 27fcd83fbcd548647a2a3f5f235c1a3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
